API for DaggerImageReconstruction

This page contains documentation of the public API of the AbstractImageReconstruction. In the Julia REPL one can access this documentation by entering the help mode with ?

Algorithm and Parameters

DaggerImageReconstruction.DaggerReconstructionAlgorithm — Type

DaggerReconstructionAlgorithm

Struct representing a Dagger-based reconstruction algorithm, which encapsulates the distrubted reconstruction execution and manages the outputs.

DaggerImageReconstruction.DaggerReconstructionParameter — Type

DaggerReconstructionParameter

Struct representing parameters for a Dagger-based reconstruction algorithm.

DaggerRecoPlan

DaggerImageReconstruction.DaggerRecoPlan — Type

DaggerRecoPlan{T}

A configuration template for an image reconstruction algorithm or parameters of type T in a different Julia process.

The DaggerRecoPlan{T} struct provides an interface similar to that of a RecoPlan, with the key distinction that setting and getting properties results in transparent data transfer to the specified remote process. DaggerRecoPlans are temporary data structures that are recreated whenever traversing the remote RecoPlan tree. They track no state, except for the reference to their remote RecoPlan counterpart.

Base.getproperty — Method

Base.getproperty(plan::DaggerRecoPlan, name::Symbol)

Get the property name of the counterpart plan. If the property is another RecoPlan, this returns a new DaggerRecoPlan pointing to the RecoPlan property.

Base.setproperty! — Method

Base.setproperty!(plan::DaggerRecoPlan, name::Symbol, x)

Set the property name of the remote plan to x. Equivalent to plan.name = x. Triggers callbacks attached to the property.

Observables.on — Method

on(f, plan::DaggerRecoPlan, property::Symbol; preprocessing = identity, kwargs...)

Registers a callback function f to be executed whenever the specified property of the given plan changes. The function f will be executed with the new value of the property on the current worker. Returns a Dagger.Chunk of the ObservableFunction. This chunk is required to unregister the listener f

To preprocess the data before it is passed to f, use the preprocessing keyword argument.

DaggerImageReconstruction.loadDaggerPlan — Function

loadDaggerPlan(filename; worker)

Load a local RecoPlan from the specified filename and interpret it on the designated worker. The resulting RecoPlan is encapsulated within a DaggerReconstructionAlgorithm.